Create Project
The Restaurant menu and Order Management System implemented in this project demonstrates effective utilization of various data structures such as vectors, queues, and classes to manage a complex system involving multiple menu categories and customer orders. Through the use of classes, we encapsulated the properties and behaviors of menu items, categories, and orders, promoting code reusability and modular design. This structure allows for easy expansion or modification of the menu, such as adding new categories or items, without affecting the core functionality of the system.